*** empty log message ***

+++ bld/nco_dst.pl
@@ -1,19 +1,20 @@
 #!/contrib/bin/perl
 				
-my $RCS_Header='$Header: /data/zender/nco_20150216/nco/bld/nco_dst.pl,v 1.11 1999-04-20 22:14:39 zender Exp $';
+my $RCS_Header='$Header: /data/zender/nco_20150216/nco/bld/nco_dst.pl,v 1.12 1999-04-27 16:11:26 zender Exp $';
 
 # Purpose: Perform NCO distributions
 
 # Usage:
 # Export tagged, public versions
 # /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln nco1_1_26
-# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --cgd_cnt nco1_1_26
-# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --wnt_cnt nco1_1_26
-# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --wnt_prs nco1_1_26
-# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --cray_prs nco1_1_26
 # /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --nst_all nco1_1_26
 # /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --acd_cnt nco1_1_26
 # /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --acd_prs nco1_1_26
+# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --cgd_cnt nco1_1_26
+# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --cray_prs nco1_1_26
+# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --dat_cnt nco1_1_26
+# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --wnt_cnt nco1_1_26
+# /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 --cln --wnt_prs nco1_1_26
 
 # Export daily snapshot
 # /home/zender/nc/nco/bld/nco_dst.pl --dbg=2 
@@ -51,23 +52,25 @@ my $True=1;
 
 my $CVSROOT='/home/zender/cvs';
 my $PVM_ARCH=$ENV{'PVM_ARCH'};
-my $RCS_Date='$Date: 1999-04-20 22:14:39 $';
-my $RCS_Id='$Id: nco_dst.pl,v 1.11 1999-04-20 22:14:39 zender Exp $';
-my $RCS_Revision='$Revision: 1.11 $';
+my $RCS_Date='$Date: 1999-04-27 16:11:26 $';
+my $RCS_Id='$Id: nco_dst.pl,v 1.12 1999-04-27 16:11:26 zender Exp $';
+my $RCS_Revision='$Revision: 1.12 $';
 my $cln=$True; # GNU standard Makefile option `clean'
 my $dbg_lvl=0;
 my $dst_cln=$False; # GNU standard Makefile option `distclean'
 my $dst_pth='/data/zender'; # Where the distribution will be exported and built
 my $main_trunk_tag='nco';
 my $nco_sng='nco';
+my $bld=$False; # Option bld; Whether to rebuild netCDF distribution
 my $nst_all=$False; # Option nst_all; Whether to install version on all machines
-my $acd_cnt=$False; # Option acd_cnt; Whether to install version on acd contrib
-my $acd_prs=$False; # Option acd_prs; Whether to install version on acd personal
-my $wnt_cnt=$False; # Option wnt_cnt; Whether to install version on winterpark contrib
-my $wnt_prs=$False; # Option wnt_prs; Whether to install version on winterpark personal
-my $cgd_cnt=$False; # Option cgd; Whether to install version on CGD contrib
-my $cgd_prs=$False; # Option cgd; Whether to install version on CGD personal
-my $cray_prs=$False; # Option cgd; Whether to install version on Cray personal
+my $acd_cnt=$False; # Option acd_cnt; Whether to install version in acd contrib
+my $acd_prs=$False; # Option acd_prs; Whether to install version in acd personal
+my $dat_cnt=$False; # Option dat_cnt; Whether to install version in dataproc contrib
+my $wnt_cnt=$False; # Option wnt_cnt; Whether to install version in winterpark contrib
+my $wnt_prs=$False; # Option wnt_prs; Whether to install version in winterpark personal
+my $cgd_cnt=$False; # Option cgd; Whether to install version in CGD contrib
+my $cgd_prs=$False; # Option cgd; Whether to install version in CGD personal
+my $cray_prs=$False; # Option cgd; Whether to install version in Cray personal
 my $vrs_tag='';
 my $www_drc='/web/web-data/cms/nco'; # WWW directory for package
 if($PVM_ARCH =~ m/SUN/){ # See Camel p. 81 for =~ and m//
@@ -90,6 +93,7 @@ $prg_vrs.='*' if length('$Locker:  $ ') > 12; # Tack '*' if it is not checked in
 
 # Parse command line arguments: '!' means Boolean, '|' is OR, '=' specifies required argument: 'i' is integer, 'f' is float
 $rcd=GetOptions(
+		'bld!' => \$bld,
 		'cln!' => \$cln,
 		'clean!' => \$cln,
 		'distclean!' => \$dst_cln,
@@ -101,6 +105,7 @@ $rcd=GetOptions(
 		'acd_prs!' => \$acd_prs,
 		'acd_cnt!' => \$acd_cnt,
 		'wnt_prs!' => \$wnt_prs,
+		'dat_cnt!' => \$dat_cnt,
 		'wnt_cnt!' => \$wnt_cnt,
 		'cray_prs!' => \$cray_prs,
 		);
@@ -110,11 +115,13 @@ if($#ARGV > 0){die "$prg_nm: ERROR Called with $#ARGV+1 positional arguments, ne
 elsif($#ARGV == 0){$vrs_tag=$ARGV[0];} # Version name is first positional argument, if present. 
 
 if($nst_all){
+    $bld=$True;
     $cgd_prs=$True;
     $cgd_cnt=$True;
     $acd_prs=$True;
     $acd_cnt=$True;
     $wnt_prs=$True;
+    $dat_cnt=$True;
     $wnt_cnt=$True;
     $cray_prs=$True;
 } # endif
@@ -123,18 +130,20 @@ if($nst_all){
 if($dbg_lvl >= 1){print ("$prg_nm: $prg_dsc, version $prg_vrs of $prg_date\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$vrs_tag = $vrs_tag\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$dbg_lvl = $dbg_lvl\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$bld = $bld\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$cln = $cln\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$dst_cln = $dst_cln\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$nst_all = $nst_all\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$mk_cmd = $mk_cmd\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$tar_cmd = $tar_cmd\n");} # endif dbg
-if($dbg_lvl >= 2){print ("$prg_nm: \$cgd_prs = $cgd_prs\n");} # endif dbg
-if($dbg_lvl >= 2){print ("$prg_nm: \$cgd_cnt = $cgd_cnt\n");} # endif dbg
-if($dbg_lvl >= 2){print ("$prg_nm: \$acd_prs = $acd_prs\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$acd_cnt = $acd_cnt\n");} # endif dbg
-if($dbg_lvl >= 2){print ("$prg_nm: \$wnt_prs = $wnt_prs\n");} # endif dbg
-if($dbg_lvl >= 2){print ("$prg_nm: \$wnt_cnt = $wnt_cnt\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$acd_prs = $acd_prs\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$cgd_cnt = $cgd_cnt\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$cgd_prs = $cgd_prs\n");} # endif dbg
 if($dbg_lvl >= 2){print ("$prg_nm: \$cray_prs = $cray_prs\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$dat_cnt = $dat_cnt\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$wnt_cnt = $wnt_cnt\n");} # endif dbg
+if($dbg_lvl >= 2){print ("$prg_nm: \$wnt_prs = $wnt_prs\n");} # endif dbg
 
 if($vrs_tag eq $main_trunk_tag || $vrs_tag eq ''){$dly_snp=$True;}else{$dly_snp=$False;}
 # NCO is distributed using the `cvs export' command, so the 
@@ -167,48 +176,50 @@ if($dbg_lvl >= 1){
     print STDOUT "$prg_nm: NCO patch version: $nco_vrs_pch\n";
 } # end if dbg 
 
-# Build distribution
-&cmd_prc("/bin/rm -r -f $dst_pth/$dst_vrs"); # Remove contents of current directory, if any
+# Build distribution from scratch
+if($bld){
+    &cmd_prc("/bin/rm -r -f $dst_pth/$dst_vrs"); # Remove contents of current directory, if any
 # NB: NCO code currently assumes -kkv and -r will fail otherwise
-if($dly_snp){
-    &cmd_prc("cvs -d $CVSROOT export -kkv -D \"1 second ago\" -d $dst_pth/$dst_vrs nco"); # Export
-}else{
-    &cmd_prc("cvs -d $CVSROOT export -kkv -r $vrs_tag -d $dst_pth/$dst_vrs nco"); # Export
-} # endelse
-&cmd_prc("printf $dst_vrs > $dst_pth/$dst_vrs/doc/VERSION"); # Stamp version in VERSION file
-
+    if($dly_snp){
+	&cmd_prc("cvs -d $CVSROOT export -kkv -D \"1 second ago\" -d $dst_pth/$dst_vrs nco"); # Export
+    }else{
+	&cmd_prc("cvs -d $CVSROOT export -kkv -r $vrs_tag -d $dst_pth/$dst_vrs nco"); # Export
+    } # endelse
+    &cmd_prc("printf $dst_vrs > $dst_pth/$dst_vrs/doc/VERSION"); # Stamp version in VERSION file
+    
 # Make sure documentation files are up to date
-my $bld_pth=$dst_pth.'/'."$dst_vrs".'/bld';
-chdir $bld_pth or die "$prg_nm: ERROR unable to chdir to $bld_pth: $!\n"; # $! is the system error sng
-&cmd_prc("$mk_cmd doc"); 
-&cmd_prc("$mk_cmd clean"); 
-
+    my $bld_pth=$dst_pth.'/'."$dst_vrs".'/bld';
+    chdir $bld_pth or die "$prg_nm: ERROR unable to chdir to $bld_pth: $!\n"; # $! is the system error sng
+    &cmd_prc("$mk_cmd doc"); 
+    &cmd_prc("$mk_cmd clean"); 
+    
 # Set up FTP server
-chdir $dst_pth or die "$prg_nm: ERROR unable to chdir to $dst_pth: $!\n"; # $! is the system error sng
-&cmd_prc("$tar_cmd -cvzf $dst_fl ./$dst_vrs"); # Create gzipped tarfile
-&cmd_prc("rsh ftp.cgd.ucar.edu /bin/rm -f /ftp/pub/zender/nco/$dst_fl"); # Remove any distribution with same name
-if($dly_snp){&cmd_prc("rsh ftp.cgd.ucar.edu /bin/rm -r -f /ftp/pub/zender/nco/nco-????????.tar.gz");} # Remove previous daily snapshots from FTP server
-&cmd_prc("rcp $dst_fl ftp.cgd.ucar.edu:/ftp/pub/zender/nco"); # Copy local tarfile to FTP server
-
+    chdir $dst_pth or die "$prg_nm: ERROR unable to chdir to $dst_pth: $!\n"; # $! is the system error sng
+    &cmd_prc("$tar_cmd -cvzf $dst_fl ./$dst_vrs"); # Create gzipped tarfile
+    &cmd_prc("rsh ftp.cgd.ucar.edu /bin/rm -f /ftp/pub/zender/nco/$dst_fl"); # Remove any distribution with same name
+    if($dly_snp){&cmd_prc("rsh ftp.cgd.ucar.edu /bin/rm -r -f /ftp/pub/zender/nco/nco-????????.tar.gz");} # Remove previous daily snapshots from FTP server
+    &cmd_prc("rcp $dst_fl ftp.cgd.ucar.edu:/ftp/pub/zender/nco"); # Copy local tarfile to FTP server
+    
 # Full release procedure (public releases only) includes update Web pages
-if(!$dly_snp){
-    &cmd_prc("rsh ftp.cgd.ucar.edu /bin/rm -f /ftp/pub/zender/nco/nco.tar.gz");
-    &cmd_prc("rsh ftp.cgd.ucar.edu \"cd /ftp/pub/zender/nco; ln -s $dst_fl nco.tar.gz\"");
-    &cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/index.shtml $www_drc/index.shtml");
-    &cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/nco.ps $www_drc/nco.ps");
-    &cmd_prc("gzip --force $www_drc/nco.ps");
-    &cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/README $www_drc/README");
+    if(!$dly_snp){
+	&cmd_prc("rsh ftp.cgd.ucar.edu /bin/rm -f /ftp/pub/zender/nco/nco.tar.gz");
+	&cmd_prc("rsh ftp.cgd.ucar.edu \"cd /ftp/pub/zender/nco; ln -s $dst_fl nco.tar.gz\"");
+	&cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/index.shtml $www_drc/index.shtml");
+	&cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/nco.ps $www_drc/nco.ps");
+	&cmd_prc("gzip --force $www_drc/nco.ps");
+	&cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/README $www_drc/README");
 #    &cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/INSTALL $www_drc/INSTALL");
-    &cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/VERSION $www_drc/VERSION");
-    &cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/ChangeLog $www_drc/ChangeLog");
-} # endif 
-
+	&cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/VERSION $www_drc/VERSION");
+	&cmd_prc("/bin/cp -f $dst_pth/$dst_vrs/doc/ChangeLog $www_drc/ChangeLog");
+    } # endif 
+    
 # Housekeeping
-if($cln){&cmd_prc("/bin/rm $dst_pth/$dst_fl");} # Remove local tarfile
-if($dst_cln){&cmd_prc("/bin/rm -r $dst_pth/$dst_vrs");} # Remove local distribution
+    if($cln){&cmd_prc("/bin/rm $dst_pth/$dst_fl");} # Remove local tarfile
+    if($dst_cln){&cmd_prc("/bin/rm -r $dst_pth/$dst_vrs");} # Remove local distribution
 
 # Sanity check
-&cmd_prc("rsh ftp.cgd.ucar.edu ls -l /ftp/pub/zender/nco");
+    &cmd_prc("rsh ftp.cgd.ucar.edu ls -l /ftp/pub/zender/nco");
+} # endif bld
 
 if($acd_prs){
     $rmt_mch='garcia.acd.ucar.edu';
@@ -241,6 +252,30 @@ if($cgd_cnt){
     &cmd_prc("rsh $rmt_mch \"cd /usr/tmp/zender/nco/bld; setenv MY_BIN_DIR /contrib/nco-1.1/bin; setenv MY_LIB_DIR /contrib/nco-1.1/lib; setenv MY_OBJ_DIR /usr/tmp/zender/nco/obj; gmake libclean binclean objclean; gmake\"");
     print STDOUT "$prg_nm: Done updating contrib NCO on $rmt_mch\n\n";
 } # endif cgd_cnt
+
+if($wnt_cnt){
+    $rmt_mch='winterpark.ucar.edu';
+#    rsh $rmt_mch 'printf $PVM_ARCH'
+    print STDOUT "\n$prg_nm: Updating contrib NCO on $rmt_mch...\n";
+    &cmd_prc("rsh $rmt_mch \"mkdir /usr/tmp/zender\"");
+    &cmd_prc("rsh $rmt_mch \"/bin/rm -r -f /usr/tmp/zender/nco*\"");
+    &cmd_prc("rcp -p ftp.cgd.ucar.edu:/ftp/pub/zender/nco/nco.tar.gz $rmt_mch:/usr/tmp/zender");
+    &cmd_prc("rsh $rmt_mch \"cd /usr/tmp/zender;tar -xvzf nco.tar.gz;rm -f nco.tar.gz;mv -f nco-* nco\"");
+    &cmd_prc("rsh $rmt_mch \"cd /usr/tmp/zender/nco/bld; setenv MY_BIN_DIR /contrib/nco-1.1/bin; setenv MY_LIB_DIR /contrib/nco-1.1/lib; setenv MY_OBJ_DIR /usr/tmp/zender/nco/obj; make libclean binclean objclean; make\"");
+    print STDOUT "$prg_nm: Done updating contrib NCO on $rmt_mch\n\n";
+} # endif wnt_cnt
+
+if($dat_cnt){
+    $rmt_mch='dataproc.ucar.edu';
+#    rsh $rmt_mch 'printf $PVM_ARCH'
+    print STDOUT "\n$prg_nm: Updating contrib NCO on $rmt_mch...\n";
+    &cmd_prc("rsh $rmt_mch \"mkdir /usr/tmp/zender\"");
+    &cmd_prc("rsh $rmt_mch \"/bin/rm -r -f /usr/tmp/zender/nco*\"");
+    &cmd_prc("rcp -p ftp.cgd.ucar.edu:/ftp/pub/zender/nco/nco.tar.gz $rmt_mch:/usr/tmp/zender");
+    &cmd_prc("rsh $rmt_mch \"cd /usr/tmp/zender;tar -xvzf nco.tar.gz;rm -f nco.tar.gz;mv -f nco-* nco\"");
+    &cmd_prc("rsh $rmt_mch \"cd /usr/tmp/zender/nco/bld; setenv MY_BIN_DIR /contrib/nco-1.1/bin; setenv MY_LIB_DIR /contrib/nco-1.1/lib; setenv MY_OBJ_DIR /usr/tmp/zender/nco/obj; make libclean binclean objclean; make\"");
+    print STDOUT "$prg_nm: Done updating contrib NCO on $rmt_mch\n\n";
+} # endif dat_cnt
 
 if($wnt_cnt){
     $rmt_mch='winterpark.ucar.edu';
